% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_aws.R
\name{process.campbell}
\alias{process.campbell}
\title{Process CAMPBELL data.}
\usage{
process.campbell(dirAWS, dirUP = NULL, upload = TRUE)
}
\arguments{
\item{dirAWS}{full path to the directory containing the AWS_DATA folder on Campbell server.\cr}

\item{dirUP}{full path to the directory containing the AWS_DATA folder on ADT server.
Default NULL, must be provided if \code{upload} is \code{TRUE}.\cr}

\item{upload}{logical, if TRUE the data will be uploaded to ADT server.}
}
\description{
Get the data from CAMPBELL database, parse and insert into ADT database.
}
